We can now fetch missing entries from bazarr

+++ src/main.rs
@@ -1,3 +1,4 @@
+use bazarr::bazarr::fetch_wanted_shows;
 use bazarr::bazarr::upload;
 use providers::errai::errai;
 use std::fs;
@@ -13,18 +14,54 @@ async fn main() {
         std::process::exit(0);
     }
 
-    let result_path = errai("Dr. Stone", "3x10", "fr").await.unwrap();
+    let wanted = fetch_wanted_shows().await;
+    match wanted {
+        Ok(shows) => {
+            if !shows.is_empty() {
+                for show in shows {
+                    if show["seriesTitle"] == "Dr. Stone" {
+                        println!("Show name: {}", show["seriesTitle"]);
+                        let wanted_ep_sonarr = show["sonarrEpisodeId"].to_string();
+                        let wanted_series_sonarr = show["sonarrSeriesId"].to_string();
+                        let wanted_series = show["seriesTitle"].to_string();
+                        let wanted_ep = show["episode_number"].to_string();
+                        for missing_sub in show["missing_subtitles"].as_array().unwrap() {
+                            let missing_lang = missing_sub["code2"].to_string();
+                            println!("Missing language: {}", missing_lang);
+                            // skip croatian
+                            if missing_lang.contains("hr") || missing_lang == "hrv" {
+                                println!("Skipping Croatian");
+                            } else {
+                                let result_path = errai(&wanted_series, &wanted_ep, &missing_lang)
+                                    .await
+                                    .unwrap();
 
-    if result_path == "" {
-        println!("Failed to download the file.");
-        std::process::exit(0);
+                                if result_path != "" {
+                                    println!("Downloaded to: {}", result_path);
+
+                                    // Then we handle uploading the file to bazarr here
+                                    upload(
+                                        &wanted_series_sonarr,
+                                        &wanted_ep_sonarr,
+                                        &missing_lang,
+                                        false,
+                                        false,
+                                        result_path.clone(),
+                                    )
+                                    .await;
+
+                                    //Then we delete the file after we're done with it
+                                    fs::remove_file(result_path).expect("Failed to delete file.");
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Err(e) => {
+            println!("Failed to fetch wanted shows: {}", e);
+            std::process::exit(1);
+        }
     }
-
-    println!("Downloaded to: {}", result_path);
-
-    // Then we handle uploading the file to bazarr here
-    upload("1", "46", "fr", false, false, result_path.clone()).await;
-
-    //Then we delete the file after we're done with it
-    fs::remove_file(result_path).expect("Failed to delete file.");
 }
